+++ Threeyes/SDK/Scripts/Component/Cursor/Controller/PostProcessing/SO/AC_SODefaultPostProcessingControllerConfig.cs
@@ -1,0 +1,6 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Cursor_Controller_PostProcessing + "DefaultPostProcessingControllerConfig", fileName = "DefaultPostProcessingControllerConfig")]
+
+public class AC_SODefaultPostProcessingControllerConfig : AC_SOConfigBase<AC_DefaultPostProcessingController.ConfigInfo> { }