+++ Threeyes/SDK/Scripts/Component/Manager/Interface/Cursor/IAC_PostProcessingManager.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface IAC_PostProcessingManager : IAC_Manager_ModInitHandler, IManagerWithController<IAC_PostProcessingController>
+{
+
+}