+++ Threeyes/SDK/Scripts/Component/Manager/Simulator/Cursor/AC_PostProcessingManagerSimulator.cs
@@ -1,0 +1,5 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class AC_PostProcessingManagerSimulator : AC_PostProcessingManagerBase<AC_PostProcessingManagerSimulator> { }